% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busy-gif.R
\name{add_busy_gif}
\alias{add_busy_gif}
\title{Busy indicator (GIF)}
\usage{
add_busy_gif(src, timeout = 100, position = c("top-right", "top-left",
  "bottom-right", "bottom-left", "full-page"), margins = c(10, 10),
  height = "50px", width = "50px")
}
\arguments{
\item{src}{Path to the GIF, an URL or a file in www/ folder.}

\item{timeout}{Number of milliseconds after the server is busy to display the spinner.}

\item{position}{Where to display the spinner: \code{'top-right'}, \code{'top-left'}, \code{'bottom-right'},
\code{'bottom-left'}, \code{'full-page'}.}

\item{margins}{Distance from margins, a vector of length two, where first element is distance from top/bottom,
second element distance from right/left.}

\item{height, width}{Height and width ot the spinner, default to \code{'50px'} for both, must be specified.}
}
\description{
Make a GIF play when server is busy and stop when idle.
}
