% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/busy-spinner.R
\name{add_busy_spinner}
\alias{add_busy_spinner}
\title{Busy indicator (spinner)}
\usage{
add_busy_spinner(spin = "double-bounce", color = "#112446",
  position = c("top-right", "top-left", "bottom-right", "bottom-left",
  "full-page"), onstart = TRUE, margins = c(10, 10), height = "50px",
  width = "50px")
}
\arguments{
\item{spin}{Style of the spinner, choice between : \code{circle}, \code{bounce}, \code{folding-cube},
\code{rotating-plane}, \code{cube-grid}, \code{fading-circle}, \code{double-bounce}, \code{dots}, \code{cube}.}

\item{color}{Color for the spinner, in a valid CSS format.}

\item{position}{Where to display the spinner: \code{'top-right'}, \code{'top-left'}, \code{'bottom-right'},
\code{'bottom-left'}, \code{'full-page'}.}

\item{onstart}{Logical, display the spinner when the application starts ?}

\item{margins}{Distance from margins, a vector of length two, where first element is distance from top/bottom,
second element distance from right/left.}

\item{height, width}{Height and width ot the spinner, default to \code{'50px'} for both, must be specified.}
}
\description{
Add a spinner in an application to show the user that 'Shiny' is busy.
}
\examples{
if (interactive()) {


}
}
