% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{html-dependencies}
\alias{html-dependencies}
\alias{html_dependency_spinkit}
\alias{html_dependency_epic}
\alias{html_dependency_shinybusy}
\alias{html_dependency_freezeframe}
\alias{html_dependency_nanobar}
\alias{html_dependency_notiflix}
\alias{html_dependency_startup}
\alias{html_dependency_loading}
\alias{html_dependency_busy}
\alias{html_dependency_notify}
\alias{html_dependency_report}
\alias{html_dependency_block}
\title{HTML dependencies used by shinybusy}
\usage{
html_dependency_spinkit()

html_dependency_epic()

html_dependency_shinybusy()

html_dependency_freezeframe()

html_dependency_nanobar()

html_dependency_notiflix()

html_dependency_startup()

html_dependency_loading()

html_dependency_busy()

html_dependency_notify()

html_dependency_report()

html_dependency_block()
}
\value{
an \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}.
}
\description{
HTML dependencies used by shinybusy
}
