% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{notify}
\alias{notify}
\title{Notifications}
\usage{
notify(
  text,
  ...,
  type = c("success", "failure", "info", "warning"),
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{text}{Text to be displayed.}

\item{...}{Options passed to JavaScript method, see https://notiflix.github.io/documentation.}

\item{type}{Type of notification: \code{success}, \code{failure}, \code{info} or \code{warning}.}

\item{session}{Default Shiny session.}
}
\value{
No value.
}
\description{
Send notifications to the user.
}
