% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinners-epic.R
\name{spin_epic}
\alias{spin_epic}
\title{Epic spinners}
\usage{
spin_epic(spin = c("flower", "pixel", "hollow-dots",
  "intersecting-circles", "orbit", "radar", "scaling-squares",
  "half-circle", "trinity-rings", "fulfilling-square"),
  color = "#112446")
}
\arguments{
\item{spin}{Name of the spinner.}

\item{color}{Color of the spinner.}
}
\value{
an HTML tag.
}
\description{
Epic spinners
}
\examples{
if (interactive()) {
  library(shiny)
  library(shinybusy)

  ui <- fluidPage(
    tags$h2("Epic spinners demo"),

    fluidRow(
      column(
        width = 2,
        tags$b("flower"),
        spin_epic(color = "#FF8000"),
        tags$br(),
        tags$b("scaling-squares"),
        spin_epic("scaling-squares", color = "#FF8000")
      ),
      column(
        width = 2,
        tags$b("pixel"),
        spin_epic("pixel", color = "#FF8000"),
        tags$br(),
        tags$b("half-circle"),
        spin_epic("half-circle", color = "#FF8000")
      ),
      column(
        width = 2,
        tags$b("hollow-dots"),
        spin_epic("hollow-dots", color = "#FF8000"),
        tags$br(),
        tags$b("trinity-rings"),
        spin_epic("trinity-rings", color = "#FF8000")
      ),
      column(
        width = 2,
        tags$b("intersecting-circles"),
        spin_epic("intersecting-circles", color = "#FF8000"),
        tags$br(),
        tags$b("fulfilling-square"),
        spin_epic("fulfilling-square", color = "#FF8000")
      ),
      column(
        width = 2,
        tags$b("orbit"),
        spin_epic("orbit", color = "#FF8000")
      ),
      column(
        width = 2,
        tags$b("radar"),
        spin_epic("radar", color = "#FF8000")
      )
    )
  )

  server <- function(input, output, session) {

  }

  shinyApp(ui, server)
}
}
